requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
Deprecated==1.2.18
frozenlist==1.5.0
future==1.0.0
h11==0.14.0
idna==3.10
line-bot-sdk==3.16.2
multidict==6.1.0
outcome==1.3.0.post0
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
